# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Application",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/feedbackComputer": {
            "get": {
                "tags": [
                    "feedback Computer"
                ],
                "summary": "Get a list of feedback",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "feedback Computer"
                ],
                "summary": "Create new feedback Computer",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "user_id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validate error"
                    },
                    "200": {
                        "description": "Success create element"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/feedbackComputer/user/{id}": {
            "get": {
                "tags": [
                    "feedback Computer"
                ],
                "summary": "Get one feedback by user id",
                "description": "",
                "operationId": "showByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/feedbackComputer/{id}": {
            "get": {
                "tags": [
                    "feedback Computer"
                ],
                "summary": "Get one feedback by  id",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "feedback id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "feedback Computer"
                ],
                "summary": "Updates a feedback Computer in the store with form data",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Feedback that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "user_id",
                                        "type": "number"
                                    },
                                    "title": {
                                        "description": "title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Error in server"
                    },
                    "400": {
                        "description": "Validate error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}